requirements.txt
anthropic>=0.45.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
schedule>=1.2.0
